Updated migrations

+++ database/migrations/2023_10_04_062030_assignment_marks.php
@@ -13,28 +13,30 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('assignment_marks', function (Blueprint $table) {
-            $table->id();
-            $table->string('sch_id');
-            $table->string('campus');
-            $table->string('period');
-            $table->string('term');
-            $table->string('session');
-            $table->unsignedBigInteger('assignment_id');
-            $table->string('student_id');
-            $table->string('subject_id');
-            $table->longText('question');
-            $table->string('question_type');
-            $table->string('question_number');
-            $table->longText('answer');
-            $table->longText('correct_answer');
-            $table->string('mark');
-            $table->string('teacher_mark');
-            $table->string('submitted');
+        if (!Schema::hasTable('assignment_marks')) {
+            Schema::create('assignment_marks', function (Blueprint $table) {
+                $table->id();
+                $table->string('sch_id');
+                $table->string('campus');
+                $table->string('period');
+                $table->string('term');
+                $table->string('session');
+                $table->unsignedBigInteger('assignment_id');
+                $table->string('student_id');
+                $table->string('subject_id');
+                $table->longText('question');
+                $table->string('question_type');
+                $table->string('question_number');
+                $table->longText('answer');
+                $table->longText('correct_answer');
+                $table->string('mark');
+                $table->string('teacher_mark');
+                $table->string('submitted');
 
-            $table->foreign('assignment_id')->references('id')->on('assignments')->onDelete('cascade');
-            $table->timestamps();
-        });
+                $table->foreign('assignment_id')->references('id')->on('assignments')->onDelete('cascade');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
