+++ database/migrations/2024_10_13_072701_create_flip_class_assessment_performances_table.php
@@ -11,24 +11,26 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('flip_class_assessment_performances', function (Blueprint $table) {
-            $table->id();
-            $table->string('sch_id');
-            $table->string('campus');
-            $table->string('period');
-            $table->string('term');
-            $table->string('session');
-            $table->unsignedBigInteger('flip_class_assessment_id');
-            $table->string('student_id');
-            $table->string('subject_id');
-            $table->string('question_type');
-            $table->string('total_mark');
-            $table->string('percentage_score');
-            $table->string('week');
+        if (!Schema::hasTable('flip_class_assessment_performances')) {
+            Schema::create('flip_class_assessment_performances', function (Blueprint $table) {
+                $table->id();
+                $table->string('sch_id');
+                $table->string('campus');
+                $table->string('period');
+                $table->string('term');
+                $table->string('session');
+                $table->unsignedBigInteger('flip_class_assessment_id');
+                $table->string('student_id');
+                $table->string('subject_id');
+                $table->string('question_type');
+                $table->string('total_mark');
+                $table->string('percentage_score');
+                $table->string('week');
 
-            $table->foreign('flip_class_assessment_id', 'fk_class_assessment_id')->references('id')->on('flip_class_assessments')->onDelete('cascade');
-            $table->timestamps();
-        });
+                $table->foreign('flip_class_assessment_id', 'fk_class_assessment_id')->references('id')->on('flip_class_assessments')->onDelete('cascade');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
